+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # WIP
 * i : Move from Mustache templates to Twig using [twig.js](https://github.com/twigjs/twig.js)
 * i : Add DNSlytics link with Page IP / Status display
+* \+ : Native messaging - accept showing notifications from native messaging
 
 # 1.0.0
 * i : Move to manifest v3, folders rework (remove data folder)